feat: implementar OrderDetail(repository, service), añadir atributos faltantes en el OrderDetailMapper(toModel)

+++ ecommerce-system/src/main/java/com/ecommercetech/orderDetail/model/OrderDetail.java
@@ -28,7 +28,7 @@ public class OrderDetail {
     private double subtotal;
 
     @Column(name = "unit_price")
-    private int unitPrice;
+    private double unitPrice;
 
     @ManyToOne
     @JoinColumn(name = "id_order")
